flask-mongo/requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
pymongo==3.12.0
Flask-Mail==0.9.1
python-dotenv==0.19.0
